added draw method for nodes and areas

+++ java/src/kit/route/a/lot/common/Coordinates.java
@@ -22,7 +22,7 @@ public class Coordinates {
         return latitude;
     }
     
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
     
@@ -30,7 +30,7 @@ public class Coordinates {
         return longitude;
     }
     
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 }
